requirements.txt
asgiref==3.5.2
black==22.6.0
certifi==2022.6.15
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.0
click==8.1.3
crispy-bootstrap5==0.6
cryptography==37.0.2
defusedxml==0.7.1
distlib==0.3.4
Django==4.0.6
django-allauth==0.51.0
django-cleanup==6.0.0
django-crispy-forms==1.14.0
django-filter==22.1
django-model-utils==4.2.0
django-notifications-hq==1.7.0
filelock==3.7.1
geographiclib==1.52
geopy==2.2.0
identify==2.5.1
idna==3.3
jsonfield==3.1.0
MarkupSafe==2.1.1
mypy-extensions==0.4.3
nodeenv==1.7.0
oauthlib==3.2.0
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
pre-commit==2.19.0
pycparser==2.21
PyJWT==2.4.0
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
swapper==1.3.0
toml==0.10.2
tomli==2.0.1
urllib3==1.26.9
virtualenv==20.15.1
